#3280 specs migrated to shared ones, and removed pending/obsolete ones
