+++ Makefile
@@ -103,7 +103,6 @@ WORKING_SPECS_6 = \
   spec/requests/api/user_layers_spec.rb \
   spec/requests/api/map_layers_spec.rb \
   spec/requests/api/records_spec.rb \
-  spec/requests/api/tables_spec.rb \
   $(NULL)
 
 WORKING_SPECS_7 = \
