+++ spec/requests/api/json/tables_controller_shared_examples.rb
@@ -12,7 +12,14 @@ shared_examples_for "tables controllers" do
       host! 'test.localhost.lan'
     end
 
+    before(:each) do
+      delete_user_data @user
+    end
+
+
     let(:params) { { :api_key => @user.api_key } }
+
+
 
     it 'returns table attributes' do
       table = create_table(
@@ -46,6 +53,95 @@ shared_examples_for "tables controllers" do
                                      description: nil,
                                      geometry_types: ["ST_MultiPolygon"]
                                      )
+      end
+    end
+
+    it "creates a new table without schema" do
+      post_json api_v1_tables_create_url(params) do |response|
+        response.status.should be_success
+        response.body[:id].should == response.headers['Location'].match(/\/([a-f\-\d]+)$/)[1]
+        response.body[:name].should match(/^untitled/)
+        response.body[:schema].should =~ default_schema
+      end
+    end
+
+    it "creates a new table without schema when a table of the same name exists on the database" do
+      create_table(name: 'untitled_table', user_id: @user.id)
+      post_json api_v1_tables_create_url(params) do |response|
+        response.status.should be_success
+        response.body[:name].should match(/^untitled_table/)
+        response.body[:name].should_not == 'untitled_table'
+        response.body[:schema].should =~ default_schema
+      end
+      @user.tables.count.should == 2
+    end
+
+    it "creates a new table specifing a name, description and a schema" do
+      post_json api_v1_tables_create_url(params.merge(
+        name: "My new blank table", 
+        schema: "code varchar, title varchar, did integer, date_prod timestamp, kind varchar", 
+        description: "Testing is awesome")) do |response|
+        response.status.should be_success
+        response.body[:name].should == "my_new_blank_table"
+        response.body[:description].should == "Testing is awesome"
+        response.body[:schema].should =~ [
+           ["cartodb_id", "number"], ["code", "string"], ["title", "string"], ["did", "number"],
+           ["the_geom", "geometry", "geometry", "geometry"],
+           ["date_prod", "date"], ["kind", "string"], ["created_at", "date"], ["updated_at", "date"]
+         ]
+      end
+    end
+
+    it "updates the metadata of an existing table" do
+      table = create_table :user_id => @user.id, :name => 'My table #1',  :tags => "tag 1, tag 2,tag 3, tag 3", :description => ""
+
+       put_json api_v1_tables_update_url(params.merge(
+          id: table.id,
+          name: table.name,
+          tags: "bars,disco", 
+          privacy: UserTable::PRIVACY_PRIVATE,
+          description: "Testing is awesome")) do |response|
+        response.status.should be_success
+        response.body[:id].should == table.id
+        response.body[:name].should == table.name
+        response.body[:privacy] == "PRIVATE"
+        response.body[:description].should == "Testing is awesome"
+        (response.body[:schema] - default_schema).should be_empty
+      end
+    end
+
+    it "updates with bad values the metadata of an existing table" do
+      table1 = create_table :user_id => @user.id, :name => 'My table #1', :tags => "tag 1, tag 2,tag 3, tag 3"
+      put_json api_v1_tables_update_url(params.merge(id: table1.id, privacy: "bad privacy value")) do |response|
+        response.status.should == 400
+        table1.reload.privacy.should == ::UserTable::PRIVACY_PRIVATE
+      end
+
+      put_json api_v1_tables_update_url(params.merge(id: table1.id, name: "")) do |response|
+        response.status.should == 400
+      end
+
+    end
+
+    it "updates a table and sets the lat and long columns" do
+      table = Table.new :privacy => UserTable::PRIVACY_PRIVATE, :name => 'Madrid Bars',
+                        :tags => 'movies, personal'
+      table.user_id = @user.id
+      table.force_schema = "name varchar, address varchar, latitude float, longitude float"
+      table.save
+      pk = table.insert_row!({:name => "Hawai", :address => "Calle de Pérez Galdós 9, Madrid, Spain", :latitude => 40.423012, :longitude => -3.699732})
+      table.insert_row!({:name => "El Estocolmo", :address => "Calle de la Palma 72, Madrid, Spain", :latitude => 40.426949, :longitude => -3.708969})
+      table.insert_row!({:name => "El Rey del Tallarín", :address => "Plaza Conde de Toreno 2, Madrid, Spain", :latitude => 40.424654, :longitude => -3.709570})
+      table.insert_row!({:name => "El Lacón", :address => "Manuel Fernández y González 8, Madrid, Spain", :latitude => 40.415113, :longitude => -3.699871})
+      table.insert_row!({:name => "El Pico", :address => "Calle Divino Pastor 12, Madrid, Spain", :latitude => 40.428198, :longitude => -3.703991})
+
+      put_json api_v1_tables_update_url(params.merge(
+        :id => table.id,
+        :latitude_column => "latitude",
+        :longitude_column => "longitude"
+      )) do |response|
+        response.status.should be_success
+        response.body[:schema].should include(["the_geom", "geometry", "geometry", "point"])
       end
     end
